+++ digest/CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `HashReader` and `HashWriter` are moved to the `digest-io` crate ([#1809])
 - `io::Write/Read` implementations in favor of the `digest_io::IoWrapper` type ([#1809])
 - `VariableOutput` trait ([#2043])
+- Implementation of `subtle::ConstantTimeEq` for `CtOutput`. Note that implementation of
+  `PartialEq`/`Eq` trait is still const time. ([#2292])
 
 [#1173]: https://github.com/RustCrypto/traits/pull/1173
 [#1334]: https://github.com/RustCrypto/traits/pull/1334
@@ -38,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1958]: https://github.com/RustCrypto/traits/pull/1958
 [#2043]: https://github.com/RustCrypto/traits/pull/2043
 [#2237]: https://github.com/RustCrypto/traits/pull/2237
+[#2292]: https://github.com/RustCrypto/traits/pull/2292
 
 ## 0.10.7 (2023-05-19)
 ### Changed